test: ProductDetailViewModel 테스트 구현
