+++ app/src/test/java/woowacourse/shopping/view/viewmodel/ProductListViewModelTest.kt
@@ -20,6 +20,7 @@ class ProductListViewModelTest {
         repository = MockProductRepository()
         viewModel = ProductListViewModel(repository)
     }
+
     @Test
     fun `offset을_기준으로_상품_리스트를_요청하면_상품_목록을_정해진_개수만큼_반환해야_한다`() {
         //given
